+++ src/lib.rs
@@ -2,14 +2,14 @@ pub mod fs;
 pub mod termio;
 
 mod client;
-mod episode;
+mod media;
 mod models;
 mod page;
 mod selectors;
 mod serie;
 
 pub use client::Client;
-pub use episode::{Episode, EpisodeID};
+pub use media::{Media, MediaType};
 pub use page::PageIterator;
 pub use serie::{Serie, SerieID};