+++ src/models/mod.rs
@@ -1,2 +1,2 @@
-pub mod episode;
 pub mod serie;
+pub mod viewer;